requirements/common.txt
Django>=5.2.7
djangorestframework>=3.16.1
django-tenants>=3.9.0
drf-nested-routers>=0.95.0
django-filter>=25.2
django-cors-headers>=4.9.0
djangorestframework-simplejwt>=5.5.1
drf-spectacular>=0.28.0
django-environ>=0.12.0

# Base de données
psycopg2-binary>=2.9.10

# Cache & Tasks
redis>=6.4.0
celery>=5.5.3
django-redis>=6.0.0

# Utils
pillow>=11.3.0
django-guardian>=3.2.0
django-auditlog>=3.3.0

# File
openpyxl>=3.1.5
pandas>=2.3.3
reportlab>=4.0